# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Food Order API",
        "description": "API documentation for the Food Order project",
        "contact": {
            "email": "zeamanuelnegussie24@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/restaurants": {
            "get": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Fetch all restaurants",
                "operationId": "0ef65917394f8eb77311d1d085ff31e4",
                "responses": {
                    "200": {
                        "description": "List of restaurants"
                    }
                }
            },
            "post": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Create a new restaurant",
                "operationId": "938db3c3288b69d111a27774178532ab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "contact_number": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "location": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Restaurant created"
                    },
                    "403": {
                        "description": "Only admins can create restaurants"
                    }
                }
            }
        },
...
        "/api/restaurants/{id}": {
            "get": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Get details of a specific restaurant",
                "operationId": "26e68feb8cf848945261ce292866ec2c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restaurant details"
                    }
                }
            },
            "put": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Update a restaurant",
                "operationId": "ab7fcec28e65e4b5fda3035e61228114",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "contact_number": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "location": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Restaurant updated"
                    },
                    "403": {
                        "description": "Unauthorized to update restaurant"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Delete a restaurant",
                "operationId": "7270982c262ae0c4e547d5eec368e436",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Restaurant deleted"
                    },
                    "403": {
                        "description": "Unauthorized to delete restaurant"
                    }
                }
            }
        },
...
    },
...
}